assistant/requirements.txt
google-generativeai==0.8.3
pydantic_settings==2.6.0
python-dotenv